requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.0
langchain-openai==0.2.1
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
